// src/modules/garmentsProduction/reports/sewing/daily-sewing-efficiency-report/daily-sewing-efficiency-report-type.ts
export type DailySewingEfficiencyReportType = {
  FLOORNAME: string;
  LINENAME: string;
  LINEINCHARGE: string;
  BUYERNAME: string;
  STYLENO: string;
  ITEMTYPE: string;
  PONO: string;
  STARTDAY: string;
  RUNNIN_DAYS: number;
  SMVSEWING: number;
  TOTALTARGET: number;
  SEWINGOUTPUT: number;
  ACTUALHOURS: number;
  RUNNINGMC: number;
  OPERATOR: number;
  HELPER: number;
  EARNINGMIN: number;
  AVAILABLEMIN: number;
  TOTALCM: number;
  TOTALFOB: number;
  AVGSMVSEWING: number;
  TOTALPRODUCTION: number;
  LOSSMIN: number;
  LOSSPICS: number;
  AVGEARNINGMIN: number;
  AVGAVAILABLEMIN: number;
  AVGLOSSMIN: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  TARGET_EARN_MIN: number;
  RUNNING_HOUR: number;
  PER_MACHINE_COST: number;
};